Add NIH SHA1 implementation mockup for porting
Mocked this up in host code to become familiar with the algorithm, and
to iron out most bugs before writing/porting to OpenCL.

For laughs, I switched sand-leek to use this implementation instead of
OpenSSL's. Performance decrease is about 90% with no change in CFLAGS,
and about 60% with -O3 instead of -O2. Of course this holds no weight, since
the implementation is only meant to be ported to OpenCL and run on graphics
cards where SIMD will skyrocket performance.