Handle multiple returned domains in  002-test-key-name.test
This patch resolves a source of unreliability in the 002-test-key-name test.
If sand-leek returned multiple "Found: " lines, then the test script would take
only the first, and expect that they RSA key is for this domain. However, this
isn't always true, due to the parallel nature of the program. Instead, this
patch adjusts the test to capture all found domains, and assert that the domain
calculated from the RSA key is one of these domains.

Tested 100% reliable over 1000 repetitions.